+++ ANSWERS.md
@@ -1,0 +1,7 @@
+#1. Node is a backend subset of javascript, it allows you to run servers using javascript.
+
+#2. Express is a web applications framework to develop web apps. Express is designed for use with node.
+
+#3. Middleware is basically used with express and it allows you to run functions at runtime
+
+#4. An endpoint allows you to connect to a website's specific location that contains a certain kind of functionality such as our apis in this assignment.
